requirements.txt
Appium-Python-Client==4.2.0
Faker==28.4.1
pandas==2.2.2
robotframework==7.1
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.6.1
selenium==4.23.0
git+https://github.com/FlamingoLindo/Utils-Scripts-For-Automation
